requirements.txt
absl-py
aiohttp
aiosqlite
async-generator
async-timeout
atomicwrites
attrs
blobfile
certifi
cffi
chardet
charset-normalizer
click
colorama
cv
DALL-E
docopt
exceptiongroup
ffmpeg-python
filelock
Flask
Flask-Cors
future
h11
idna
imageio
iniconfig
itsdangerous
jax
Jinja2
MarkupSafe
multidict
mypy
mypy-extensions
numpy
opt-einsum
outcome
packaging
Pillow
pipreqs
pluggy
py
py-cord 
pycparser
pycryptodomex
pygame
PyNaCl
pyparsing
PySocks
pytest
python-dotenv
requests
scipy
selenium
six
sniffio
sortedcontainers
tk
tomli
torch
torchvision
trio
trio-websocket
typing_extensions
urllib3
Werkzeug
wsproto
xmltodict
yarg
yarl
youtube-dl
